work on the base (edits)

+++ app/Http/Controllers/ResumesController.php
@@ -7,23 +7,23 @@ use App\Models\Resume;
 use App\Models\Level;
 use App\Models\User;
 use App\Models\Profile;
-use App\Models\Addresse;
+use App\Models\Address;
 use App\Models\Language;
 use App\Models\Skill;
 use App\Models\Education;
-use App\Models\Work;
-use App\Models\Hobbi;
+use App\Models\Experience;
+use App\Models\Hobby;
 use App\Models\Field;
 use Illuminate\Support\Facades\Validator;
 use PDF;
 
 class ResumesController extends Controller
 {
-    public function getPdf(Request $request, $slag)
+    public function getPdf(Request $request, $slug)
     {
         $user = auth()->user();
-        $slag = $request->slag;
-        $resume = Resume::slag($slag);
+        $slug = $request->slug;
+        $resume = Resume::slug($slug);
         $data = [
             'user' => $user,
             'resume' => $resume,
@@ -52,24 +52,22 @@ class ResumesController extends Controller
     {
         $user = auth()->user();
         $resume = $request->resumeform;
-        $slug = $resume['slag'];
+        $slug = $resume['slug'];
         if(count($user->resumes) < 3 || !isset($user->resumes)) {
             $user_id = $user->id;
 
             $resume += ['user_id' => $user_id];
             $resume_data = Validator::make($resume, [
                 'user_id' => ['required'],
-                'slag' => ['string', 'min:3', 'max:255'],
+                'slug' => ['string', 'min:3', 'max:255'],
                 'job_title' => ['string', 'min:3', 'max:100'],
                 'description' => ['string', 'min:3', 'max:10000']
             ]);
             if ($resume_data->fails()){
-                return response()->json(['error' => 'No valid form resume!'], 409);
+                return response()->json(['error' => 'No valid form resume!'], 415);
             }
-            else {
-                Resume::create($resume);
-                return response()->json(['success' => 'Resume '.$slug.' user`s '.$user->name.' created!'], 201);
-            }
+            Resume::create($resume);
+            return response()->json(['success' => 'Resume '.$slug.' user`s '.$user->name.' created!'], 201);
         }
         else{
             return response()->json(['error' => 'Resume '.$slug.' user`s '.$user->name.' no created! You already have three resumes!'], 409);
@@ -78,14 +76,14 @@ class ResumesController extends Controller
 
     public function getResume(Request $request)
     {
-        $slag = $request->slag;
-        $resume = Resume::slag($slag);
-        return response()->json(['slag' => $resume->slag, 'job_title' => $resume->job_title, 'description' => $resume->description], 200);
+        $slug = $request->slug;
+        $resume = Resume::slug($slug);
+        return response()->json(['slug' => $resume->slug, 'job_title' => $resume->job_title, 'description' => $resume->description], 200);
     }
 
-    public function publicShowResume(Request $request, $slag)
+    public function publicShowResume(Request $request, $slug)
     {
-        $resume = Resume::slag($slag);
+        $resume = Resume::slug($slug);
         $user = $resume->user;
         return view('resume', ['resume' => $resume, 'user' => $user]);
     }
@@ -97,29 +95,24 @@ class ResumesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $slag)
+    public function update(Request $request, $slug)
     {
-        // dd($slag);
-        $resume = Resume::slag($slag);
-        // dd($slag);
+        $resume = Resume::slug($slug);
         $update_resume['job_title'] = $request->resumeform['job_title'];
         $update_resume['description'] = $request->resumeform['description'];
         $update_resume['user_id'] = $resume->user_id;
-        $update_resume['slag'] = $resume->slag;
-        // dd($resume);
+        $update_resume['slug'] = $resume->slug;
         $resume_data = Validator::make($update_resume, [
             'user_id' => ['required'],
-            'slag' => ['string', 'min:3', 'max:255'],
+            'slug' => ['string', 'min:3', 'max:255'],
             'job_title' => ['string', 'min:3', 'max:100'],
             'description' => ['string', 'min:3', 'max:10000']
         ]);
         if ($resume_data->fails()){
-                return response()->json(['error' => 'No valid form resume!'], 409);
+                return response()->json(['error' => 'No valid form resume!'], 415);
             }
-            else {
-                $resume->update($update_resume);
-                return response()->json(['success' => 'Resume '.$resume->slag.' update!'], 201);
-            }
+        $resume->update($update_resume);
+        return response()->json(['success' => 'Resume '.$resume->slug.' update!'], 201);
     }
 
     /**
@@ -130,8 +123,14 @@ class ResumesController extends Controller
      */
     public function destroy(Resume $resume)
     {
-        $resume->delete();
-        $resumes = auth()->user()->resumes;
-        return $resumes;
+        if ($resume->user->id == auth()->user()->id) {
+            $resume->delete();
+            $resumes = auth()->user()->resumes;
+            return $resumes;
+        }
+        else {
+            return response()->json(['error' => 'Unauthorized!'], 401);
+        }
+
     }
 }
